Rename assemblies to CRCTray
