+++ CRCTray.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29911.84
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tray-windows", "tray-windows.csproj", "{13AFC4D0-BA4B-437E-8994-D57B61F915B8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CRCTray", "CRCTray.csproj", "{13AFC4D0-BA4B-437E-8994-D57B61F915B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
