+++ Forms/StatusForm.cs
@@ -3,9 +3,9 @@ using System.IO;
 using System.Drawing;
 using System.Windows.Forms;
 using System.Threading.Tasks;
-using tray_windows.Communication;
+using CRCTray.Communication;
 
-namespace tray_windows
+namespace CRCTray
 {
     public partial class StatusForm : Form
     {
