+++ Helpers/DisplayMessageBox.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Forms;
 
-namespace tray_windows
+namespace CRCTray.Helpers
 {
     class DisplayMessageBox
     {
