+++ appveyor.yml
@@ -5,10 +5,10 @@ image: Visual Studio 2019
 configuration: Release
 
 before_build:
-  ps: nuget restore tray-windows.sln
+  ps: nuget restore CRCTray.sln
 
 build:
-  project: tray-windows.sln
+  project: CRCTray.sln
 
 after_build:
   ps: ls bin\
